FIX: restructuring the project files
