+++ src/_posts/body/running/running_title/content.mdx
@@ -1,11 +1,11 @@
 ---
 title: "title running_title_1"
 date: 2024-11-01
-description: description running_title_1
+description: description running_1
 keywords: [keyword1,keyword2,keyword3]
 thumbnail: /posts/nextjs_blog/setup/thumbnail.png
 ---
  
-#### 환영 running_title_1
+#### 환영 running_1
  
 안녕하세요. 이 글은 mdx파일로 작성되었습니다. running_title_1
