+++ src/_posts/moral/book/book_title/content.mdx
@@ -1,11 +1,11 @@
 ---
 title: "title book_title1"
 date: 2024-11-01
-description: description book_title1
+description: description book 1
 keywords: [keyword1,keyword2,keyword3]
 thumbnail: /posts/nextjs_blog/setup/thumbnail.png
 ---
  
-#### 환영 book_title
+#### 환영 book1
  
 안녕하세요. 이 글은 mdx파일로 작성되었습니다. 1
