--- src/_posts/moral/stock/stock_title_2/content.mdx
@@ -1,11 +1,0 @@
----
-title: "title stock_title_2"
-date: 2024-11-01
-description: description 테스트stock_title_2
-keywords: [keyword1,keyword2,keyword3]
-thumbnail: /posts/nextjs_blog/setup/thumbnail.png
----
- 
-#### 환영 stock_title_2
- 
-안녕하세요. 이 글은 mdx파일로 작성되었습니다. stock_title_2
