--- src/components/etc/ServerButton.tsx
@@ -1,12 +1,0 @@
-import Link from "next/link";
-
-const ServerButton = () => {
-  return (
-    <div>
-      <div className="w-36 bg-blue-100">ServerButton</div>
-      <Link href={"/test1"}>test1 page</Link>
-    </div>
-  );
-};
-
-export default ServerButton;
