--- src/config/info.ts
@@ -1,1 +1,0 @@
-export const BASE_URL = "https://seunghowhite.github.io";
